Pass window values via signal
Direct writing of the TunnelWindow items wasn't working reliably (window freezes, sometimes crasshes). Changed code to use a signal to pass a ProcessedSample to a slot in TunnelWindow, which then unpacks it and writes it to the items in the window as the events wake up.

Discard the text Spinner, as it should no longer be needed to note if the GUI is updating properly.