Move access to persist to SampleCollector.
Trying to write persist.ini from TunnelMain and SampleCollector is a Bad Thing.

Concentrate read/writes on persist.ini within SampleCollector and provide hooks to TunnelMain to retrieve and set Run Name and Configuration.

The scope of SampleCollector is creeping, and suggests that there may need to be another object that handles persist.ini specifically in the future.