Add platform AoA offset calculation.
Add a AoA platform offset value to persist.ini.

When changed in the calibration dialog, update the AoA offset and copy back to persist.ini. Then in SampleCollector. add the saved offset value to both platform AoA and wing AoA.

Update and display the platform AoA offset value on the main window.